ReadMe.md
Tower Of Hanoi Game

The game is quite simple.  You have 3 different stacks  and blocks ranging  from smallest to biggest on 1 stack. 
the Object of the game is to ultimately move all blocks from 1 stack to the last stack in the order of biggest on the bottome
and smallest on top.

The only rules are: 1) move 1 block at a time, 2) you can't place a bigger block on top of a smaller block. Viola...that's it.